// src/app/shared/shared/enum/common/type-control.enum.ts
export enum TypeControl {
        /// <summary>
        /// mặc định hoặc 1 dòng
        /// </summary>
        DefaultType = 0,
        /// <summary>
        /// Loại nhập 1 dòng
        /// </summary>
        OneRow = 1,
        /// <summary>
        /// Loại nhập nhiều dòng
        /// </summary>
        MultiRow = 2,
        /// <summary>
        /// Loại combobox chọn một
        /// </summary>
        Combobox = 3,
        /// <summary>
        /// Loại combobox chọn nhiều
        /// </summary>
        MultiCombobox = 4,
        /// <summary>
        /// Loại ngày
        /// </summary>
        Date = 5,
        /// <summary>
        /// Loại ngày tháng
        /// </summary>
        DateTime = 6,
        /// <summary>
        /// Loại Tháng năm
        /// </summary>
        MonthYear = 7,
        /// <summary>
        /// Loại số
        /// </summary>
        Number = 8,
        /// <summary>
        /// Loại mã  tự tăng
        /// </summary>
        AutoNumber = 9,
        /// <summary>
        /// Loại tiền
        /// </summary>
        Currency = 10,
        /// <summary>
        /// Loại link
        /// </summary>
        Hyperlink = 11,
        /// <summary>
        /// Loại tính chọn
        /// </summary>
        Checkbox = 12,
        /// <summary>
        /// Loại chọn ảnh
        /// </summary>
        UploadImage = 13,
        /// <summary>
        /// Loại chọn tài liệu
        /// </summary>
        UploadDocument = 14,
        /// <summary>
        /// Tỷ lệ
        /// </summary>
        Percent = 15,

        /// <summary>
        /// Dạng combo tree
        /// </summary>
        ComboTree = 16,

        /// <summary>
        /// Control dạng công thức
        /// </summary>
        Formula = 17,
        /// <summary>
        /// Loại dữ liệu email
        /// </summary>
        Email = 18,
        /// <summary>
        /// Loại dữ liệu thập phân
        /// </summary>
        Decimal = 19,
        /// <summary>
        ///  kiểu control chọn người
        /// </summary>
        SelectHuman = 20,
        /// <summary>
        /// Control soạn thảo văn bản
        /// </summary>
        Editor = 21,
        /// <summary>
        /// Kiểu control năm
        /// </summary>
        Year = 22,
        /// <summary>
        /// Kiểu tree
        /// </summary>
        TreeBox = 23,
        /// <summary>
        /// Kiểu tree chọn nhiều
        /// </summary>
        TreeBoxMulti = 24,

        EnumType = 25,
        /// <summary>
        /// kiểu hiển thị dung lượng
        /// </summary>
        FileSize = 26,

        /// <summary>
        /// loại gợi ý
        /// </summary>
        Autocomplete = 27,

}